+++ Example/ReusableDemo iOS/CollectionViewController.swift
@@ -21,7 +21,7 @@ final class CollectionViewController: UICollectionViewController {
     //    self.collectionView.registerReusableCell(MyStoryBoardIndexPathCell)
 
     collectionView.register(supplementaryViewType: CollectionHeaderView.self,
-                            ofKind: UICollectionElementKindSectionHeader)
+                            ofKind: UICollectionView.elementKindSectionHeader)
 
     if let flowLayout = self.collectionView?.collectionViewLayout as? UICollectionViewFlowLayout {
       flowLayout.headerReferenceSize = CGSize(width: collectionView.bounds.size.width, height: 60)
@@ -67,3 +67,10 @@ final class CollectionViewController: UICollectionViewController {
       }
   }
 }
+
+/// Swift < 4.2 support
+#if !(swift(>=4.2))
+private extension UICollectionView {
+  static let elementKindSectionHeader = UICollectionElementKindSectionHeader
+}
+#endif