+++ Sources/View/NibOwnerLoadable.swift
@@ -37,21 +37,25 @@ public extension NibOwnerLoadable where Self: UIView {
    Adds content loaded from the nib to the end of the receiver's list of subviews and adds constraints automatically.
    */
   func loadNibContent() {
-    let layoutAttributes: [NSLayoutAttribute] = [.top, .leading, .bottom, .trailing]
-    for view in Self.nib.instantiate(withOwner: self, options: nil) {
-      if let view = view as? UIView {
-        view.translatesAutoresizingMaskIntoConstraints = false
-        self.addSubview(view)
-        layoutAttributes.forEach { attribute in
-          self.addConstraint(NSLayoutConstraint(item: view,
-            attribute: attribute,
-            relatedBy: .equal,
-            toItem: self,
-            attribute: attribute,
-            multiplier: 1,
-            constant: 0.0))
-        }
-      }
+    let layoutAttributes: [NSLayoutConstraint.Attribute] = [.top, .leading, .bottom, .trailing]
+    for case let view as UIView in Self.nib.instantiate(withOwner: self, options: nil) {
+      view.translatesAutoresizingMaskIntoConstraints = false
+      self.addSubview(view)
+      NSLayoutConstraint.activate(layoutAttributes.map { attribute in
+        NSLayoutConstraint(
+          item: view, attribute: attribute,
+          relatedBy: .equal,
+          toItem: self, attribute: attribute,
+          multiplier: 1, constant: 0.0
+        )
+      })
     }
   }
 }
+
+/// Swift < 4.2 support
+#if !(swift(>=4.2))
+private extension NSLayoutConstraint {
+  typealias Attribute = NSLayoutAttribute
+}
+#endif